Add extract-match mixin
Works similar to apply-match but exports the @__match variable to the
parent scope instead of accepting a ruleset. I have a suspicion that
this will generally be easier to wield and produce more predictable
results than all the crazy extra scoping bullshit I always have to try
and do when using the "closure"-ish approach I use with apply-match.