Add basic background color generator
Supports specifying background colors as an array of colors, or an
object where the label is what you want to name that color as a
background utility, and the value is the color to use.

Everything can be specified as either camelCase or kebab-case so you
don't need to quote the shit out of everything but classes will always
be rendered as kebab-case.